requirements.txt
requests
supabase
python-dotenv
pandas
streamlit
plotly